Send lidar distance as float, not uint16_t

+++ program/system/fcb_link.h
@@ -9,7 +9,7 @@
 	uart2_puts(buffer, size); \
 
 
-void send_flow_to_fcb(uint16_t *lidar_distance, float *lidar_velocity,
+void send_flow_to_fcb(float *lidar_distance, float *lidar_velocity,
                       float *flow_vx, float *flow_vy,
                       float *accel_x, float *accel_y, float *accel_z,
                       float *time, float *period, float *frequency, float *quality);
